Crate graph_algorithms

Modules

list_of_algorithms 🔒

Functions

bellmanford

Performs Bellmanford algorithm on a given weighted graph. Prints a graph with shortest distance from one vertex to another vertex.

bfs

Performs Breadth First Search algorithm on a given graph represented as an adjacency list. Prints a graph, where each inner vector contains the nodes of a strongly connected component in sorted order.

dfs

Performs Depth first search algorithm on a given directed graph represented as an adjacency list. Prints a graph, where the nodes of the graph are in the visited order.

dijkstras

Performs Dijkstra’s algorithm on a given directed graph represented as an adjacency list. Prints a vector of vectors, where each inner vector contains the nodes of dijkstras in sorted order.

kosaraju

Performs Kosaraju’s algorithm on a given directed graph represented as an adjacency list. Prints a vector of vectors, where each inner vector contains the nodes of a strongly connected component in sorted order.

main 🔒

The main function displays a list of different algorithms and allows the user to select one.