Module graph_algorithms::list_of_algorithms::kosaraju::kosaraju

Functions

kosaraju

Performs Kosaraju’s algorithm on a given directed graph represented as an adjacency list. Prints a vector of vectors, where each inner vector contains the nodes of a strongly connected component in sorted order.

kosaraju_algorithm

The Kosaraju’s algorithm is used to find strongly connected components. Given a directed graph represented as an adjacency list (Vec[Vec[]]), returns a vector of strongly connected components.